Update 08-projects.Rmd
At least 4 students during office hours on Monday did not catch the fact that in question # 1 we did not want them to use the 'mean' function. They thought that the first 'mean' stood for the average hence a lot of them used the mean function to solve that question. I made a small edit in order to distinguish the use of the 'mean' function in R and the word mean (average). Hopefully that would avoid the confusion.