[Stat 1900- project 13] Edits
Dear @kevinamstutz  please find below the edits I suggest we make for project 13.

Question # 1: For this question we might want to explain the code and output displayed in the hint section. It seems that we are comparing how fast `read.csv` and `fread` work but no information was provided before to let the student now that it is what is being done. I therefore added a small sentence to clarify what the code and output  was about.
Also the code provided reads in the first 100000 rows of the dataset students might think that we want them to only read in this many rows from `reviews` as well.
I have changed some of the wording in the hint section. 
I added a link for `fread` and will update the R documentation to include an example on how to used it in a different pull request.

Question # 2: Fixed the relevant topic link for `%in%`.

Question # 3: changed the position of the link for (...) in order to make it more visible to students in the example book. 
fixed the relevant topic link for %in% 
In the question it is not clear whether we want the average `n_types` for IN and MI separately or  as one mean together.

Question # 4: This question was not clear to me I tried to reword it in the way I thought would make it clear to the students.
I fixed the links in the relevant topic section

Question # 5: fixed a typo in the hint section and fixed the links in the relevant topic section. Fixed typos in the item to submit box