[Stat 1900 - project 14] Edits
Question # 1: Added the link for ` the function `fread` in the relevant topic section.
Added a link for the function `setdiff` and will be adding a section in the R section of the example book using another pull request.
 Also fixed small grammatical errors within the question's paragraph.

Question # 2: Added the link for the function `in` in the relevant topic section. Some minor sentence changes.

Question # 3: Added a link for the function `intersect` and will be adding a section in the R section of the example book.

Question # 4: Added the link for the functions `in`  and  `setdiff` in the relevant topic section. Reworded question sentence.

Question # 5:  Added the link for the functions  `intersect` in relevant topics.

Question # 6: NA