STAT 19000 Project 2: Edits
Dear @kevinamstutz , I found very few edits to make on this project.
Question 1: NA
Question 2: NA
Question 3: I rephrased part of the question to reflect the way I understood it. I am not sure whether that was what was meant or not.
Question 4: I personally don't see a big difference in the shape of the plot with or without the last 50 values. The y-axis scale changes but the shape of the plot remains the same.
Question 5: NA