STAT 19000 Project 3 edits
Question 1 : added a link to the python head() function within the example book.
Question 2 :  NA
Question 3 : added a direct link to the matplotlib barplot.
Question 4 :  Slightly changed the wording of the last sentence in the question to make it obvious to students that they should recode all three examples.
Question 5 :  Part 1 -- I tried this question and for constructing the dictionary I could not get the code to work without using the setdefault() which we have not introduced to the student. I am not sure if I am the one who cant think of a way of not using this function but I tried without it and it would let me append values to a specific key in the dictionary otherwise.  Here is the code I worked with 
` # for each year (key),long and lat 
mystates = myDF['state'].dropna().to_list() # liste of states
unique_states = list(set(mystates)) # unique list of states
geoDict={}
for i in range(0,len(states_list)):
    geoDict.setdefault(states_list[i][0], [])
    geoDict[states_list[i][0]].append(tuple(states_list[i])[1:3])`
Part 2-- I could not get the map to render I get the following message: DriverError: /class/datamine/data/craigslist/cb_2018_us_state_20m.shp: Permission denied
Question 6 : Fixed small typo