STAT 19000 Project 7 : Review
Question 1 :  NA

Question 2:  made change to file name to reflect the fact that it should be **businesses.parquet** and not **business.parquet**.
Question 3: For the first part of this question ( the type of the values) I am not sure what the answer should be I get `dict_value` as the type,  but looking at the values I see that some of them are actually dictionaries themselves.
Changed the wording in the second paragraph of the question to reflect the fact that  the result of the function should be a file called `new_businesses.parquet` saved in the output path.
Removed the 'f' that was in front of the output_dir.
For this question after visiting the stackoverflow page linked I could not get the codes using the apply function to work. I had to combine the suggestions of two different contributors using the `pd.json_normalize`  function to get the code to work. Is that expected of students or do you have a specific contributor answer you hope they will find that will work for this project ?
question 4: NA 
question 5: NA