[Stat 19000 ]project 8 review
Question 1:  Added a couple of words to second sentence to clarify the fact that we want to return the information of `user_id`'s friends. 
The link provided in  the `as well as **this** page` points back to the beginning of the project page. 

Question 2:  Should we provide students with a business_id to test their function, since the submission box asks for the output of running the code.
Here is the test I  used and its output in case you think we should add it 
print(calculate_avg_business_stars("f9NumwFMBDn751xgFiRbNA")) # 3.1025641025641026
I added it in the project. Feel free to remove it if you do not think it should be added.

I thought we should provide them some information on how groupby works and other relevant topic so I added a list of relevant topic along with their links.

Question 3 - 6: added a list of relevant topics.