STAT 19000 Project 9 review
Question 1: fixed a small typo. 
I do not understand the hint provided. Why do the students need a percentage ?\

Question 2: Added the word `train` to clarify the reference made later in the paragraph.
I do not believe the Stat 19000 students were introduced to the tidyverse package last year. I think it would be better to first show the tapply function example before the tidyverse one so I switched their position. 
Also, I am not sure about the need to mention the  `cosine_similarity_matrix` for this question.

Question 3-4: Should we ask them to print the head and/or shape of the resulting DataFrame ? I have added that in the item to submit box.