Stat 19000 Project 12 Edits
Dear @kevinamstutz here are my comments for this project.

There is an incomplete sentence at the end of the *Dataset* section.

I only have a general comment for this project. I think students will need more guidance for this project as the concept of classes is very new and unlike what they have been doing so far. They will either need Dr. ward videos or links with examples. The links provided for the dunder methods provide a general description of the methods with no example for them to refer to. I would also like to request that if the solution to the project are ready that they be posted earlier for the TAs as I foresee students asking questions as soon as Saturday before Dr. Ward is able to upload his example videos.
Thank you