[INS 199 Drew Sukova] created data-frames.adoc from R content (#299)
* Update data-frames.adoc

Moved the old Examples Book content to the new Examples Book, making minor revisions for flow.

* Update data-frames.adoc

Updated the data frames adoc with all of the requested changes. On another note, I looked in the examples book asciidoc tutorial and did a quick Google search, through neither of which I could find how to "hide" solutions/examples with a drop-down arrow. If you let me know how I will add that in future. Thanks!

* Update data-frames.adoc

Switched the placement of the headers - now questions can be linked to instead of their solutions.

* Revised data-frames.adoc for the data.frame page

Removed the unneeded bold font and added newlines. For the newlines, I used {sp} + and grouped some text and headers for visual flow. Let me know if there's a better way of doing newlines or if you need headers explicitly separated (I didn't know if there was a particular reason so I grouped it by looks instead). Thanks!

* Finalized data-frames.adoc

Updated with suggestions from @naomim01

* Update data-frames.adoc

Removing === around code blocks to align with the rest of the book.

Co-authored-by: David Glass <dglass@purdue.edu>