Added note for Jackson's OH
Dr. Sharples approved totally virtual office hours for Jackson so that he can still do his normal workload in-office. This note reflects that. If students feel the need to meet in person or Jackson's hours don't work for them as is, his email has been included so they can schedule that with him.