Projects Appendix Restructuring (#995)
tldr: Projects appendix was extremely unmodular and broke anytime projects were moved from current-projects to project-archive. The new structure is completely modular and means that the only difference between a project in the archive and a project in the current projects is the nav.adoc file. Moving a project from the current projects to the archive is as simple as changing navigation and nothing else.

* Projects Appendix Restructuring

Let's just see how badly this breaks navigation and go from there, yeah?

* Small nav bar fixes

And so it begins, the wonderful fixing of navigation. Hopefully this takes less that 30 hours to complete

* More nav changes

* More nav bar testing

* More nav testing

I really need to start learning how to do this regularly, these nav bars always give me so much trouble

* MORE nav fixes

* Think Nav is Fixed

* Seriously Nav Fixes

Actually no idea how this stuff works

* Fixed? Nav

* Broke build pipeline?

Build attempt immediately failed. Very strange behavior

* Nav fix attempt

* Big Nav Fixes

This should fix most of the issues and put me in good position to start merging branches

* Change names to Valid CRNs

Per Dr. Ward request to have course numbers of the format XXX00 and not XXX.