Project 11 Rewrite Complete
I think this is a good way to cap off this project. Again, I will need some review from others to make sure I didn't overdo it on the difficulty but I _think_ that this is veeeery doable for a freshman given the amount of sample code they've been given.