Test (#1014)
* Test

* Sprint #1 Updates

First pass

* Sprint #2 updates

Changes still need to be made to the resume and mock interview page. Options need to be updated for Purdue Students. We might want to also include screenshots of how to navigate to the CCO.

* Sprint #3 updates

Minor changes also made to Sprint #2 and resume/mock interview page

* Sprint #4 Updates

Made minor fixes to Sprint #3

* Sprint #5 updates

With minor changes to Sprint #4

* Sprint #6 and #7 updates

* Minor fix Sprint #6

* Article added for Kanban Board Intro

* Team contract updates

Still need to create an outline.

* Update TA information

* PD updates for TA

Also created page for updating the project charter

* Final Presentation updates

* Final presentation updates 2.0 and minor tweaks to other TA pages